requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
google-genai
google-generativeai
python-dotenv
sqlalchemy
asyncpg
httpx
python-jose[cryptography]
